+++ R/globals.R
@@ -1,1 +1,1 @@
-utils::globalVariables(c("lib_sizes"))
+utils::globalVariables(c("libsizes","x_xmap_y_mean", "y_xmap_x_mean"))