+++ R/variable_check.R
@@ -1,0 +1,7 @@
+.check_character = \(cause,effect){
+  if (!inherits(cause,"character") || !inherits(effect,"character")) {
+    stop("The `cause` and `effect` must be character.")
+  }
+  varname = c(cause,effect)
+  return(varname)
+}