+++ R/zzz.R
@@ -1,3 +1,5 @@
 .onLoad = function(...) {
   loadNamespace("Rcpp")
+  loadNamespace("sf")
+  loadNamespace("terra")
 }