+++ NEWS.md
@@ -1,6 +1,16 @@
+# spEDM 1.2
+
+* Encapsulate the `gccm` function using the S4 class (#72).
+
+* Add options for `tau`, `k`, and `progressbar` in `gccm` (#69).
+
+* Add `print` and `plot` s3 methods for `gccm` result (#64).
+
+* Require `sdsfun` package version `0.7.0` or higher (#61).
+
 # spEDM 1.1
 
-* Resolve CRAN autocheck issues, no significant API changes.
+* Resolve CRAN auto check issues, no significant API changes.
 
 # spEDM 1.0